+++ src/main/java/com/myproyect/springboot/repos/FabricaGaussRepository.java
@@ -1,10 +1,10 @@
 package com.myproyect.springboot.repos;
 
-import com.myproyect.springboot.domain.concurrency.FabricaGauss;
+import com.myproyect.springboot.domain.factory.FabricaGauss;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FabricaGaussRepository extends JpaRepository<FabricaGauss, Long> {
+public interface FabricaGaussRepository extends JpaRepository<FabricaGauss, Integer> {
 }
