+++ src/main/java/com/myproyect/springboot/rest/MaquinaResource.java
@@ -27,20 +27,20 @@ public class MaquinaResource {
 
     @GetMapping("/{id}")
     @ApiResponse(responseCode = "200", description = "Get a maquina by ID")
-    public ResponseEntity<MaquinaDTO> getMaquina(@PathVariable final Long id) {
+    public ResponseEntity<MaquinaDTO> getMaquina(@PathVariable final Integer id) {
         return ResponseEntity.ok(maquinaService.get(id));
     }
 
     @PostMapping
     @ApiResponse(responseCode = "201", description = "Create a new maquina")
-    public ResponseEntity<Long> createMaquina(@RequestBody final MaquinaDTO maquinaDTO) {
-        final Long createdId = maquinaService.create(maquinaDTO);
+    public ResponseEntity<Integer> createMaquina(@RequestBody final MaquinaDTO maquinaDTO) {
+        final Integer createdId = maquinaService.create(maquinaDTO);
         return new ResponseEntity<>(createdId, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
     @ApiResponse(responseCode = "200", description = "Update an existing maquina")
-    public ResponseEntity<Void> updateMaquina(@PathVariable final Long id,
+    public ResponseEntity<Void> updateMaquina(@PathVariable final Integer id,
                                               @RequestBody final MaquinaDTO maquinaDTO) {
         maquinaService.update(id, maquinaDTO);
         return ResponseEntity.ok().build();
@@ -48,7 +48,7 @@ public class MaquinaResource {
 
     @DeleteMapping("/{id}")
     @ApiResponse(responseCode = "204", description = "Delete a maquina")
-    public ResponseEntity<Void> deleteMaquina(@PathVariable final Long id) {
+    public ResponseEntity<Void> deleteMaquina(@PathVariable final Integer id) {
         maquinaService.delete(id);
         return ResponseEntity.noContent().build();
     }
